+++ typescript/ui/src/core/components/ChatMessage.tsx
@@ -1,16 +1,21 @@
 import { forwardRef, useRef, useState } from "react";
 import { Button, cn } from "@heroui/react";
+import { Icon } from "@iconify/react";
 import Markdown from "react-markdown";
 import remarkGfm from "remark-gfm";
-import type { ChatMessage } from "../../types/history.ts";
-import { ConfigResponse, FormType, MessageRole } from "../../types/api";
-import { Icon } from "@iconify/react";
-import DelayedTooltip from "./DelayedTooltip";
+
+import { ChatMessage } from "../../types/history.ts";
+import { ConfigResponse, FormType, MessageRole } from "../../types/api.ts";
 import { useThemeContext } from "../../contexts/ThemeContext/useThemeContext.ts";
 import { Theme } from "../../contexts/ThemeContext/ThemeContext.ts";
+import DelayedTooltip from "./DelayedTooltip";
 
 export type ChatMessageProps = {
-  classNames?: string[];
+  classNames?: {
+    wrapper?: string;
+    innerWrapper?: string;
+    content?: string;
+  };
   chatMessage: ChatMessage;
   onOpenFeedbackForm?: (id: string, name: FormType) => void;
   likeForm: ConfigResponse[FormType.LIKE];
@@ -53,7 +58,7 @@ const ChatMessage = forwardRef<HTMLDivElement, ChatMessageProps>(
         className={cn(
           "flex gap-3",
           { "flex-row-reverse": rightAlign },
-          ...(classNames ? classNames : []),
+          classNames?.wrapper,
         )}
       >
         <div
@@ -69,6 +74,7 @@ const ChatMessage = forwardRef<HTMLDivElement, ChatMessageProps>(
             className={cn(
               "relative rounded-medium px-4 py-3 text-default",
               rightAlign && "bg-default-100",
+              classNames?.innerWrapper,
             )}
           >
             {rightAlign ? (
@@ -76,6 +82,7 @@ const ChatMessage = forwardRef<HTMLDivElement, ChatMessageProps>(
                 className={cn(
                   "prose whitespace-pre-line",
                   theme === Theme.DARK && "dark:prose-invert",
+                  classNames?.content,
                 )}
               >
                 {content}
@@ -86,6 +93,7 @@ const ChatMessage = forwardRef<HTMLDivElement, ChatMessageProps>(
                   className={cn(
                     "markdown-container prose max-w-full",
                     theme === Theme.DARK && "dark:prose-invert",
+                    classNames?.content,
                   )}
                   remarkPlugins={[remarkGfm]}
                 >
@@ -129,7 +137,7 @@ const ChatMessage = forwardRef<HTMLDivElement, ChatMessageProps>(
                             isIconOnly
                             variant="ghost"
                             className="p-0"
-                            aria-label="Like message"
+                            aria-label="Rate message as helpful"
                             onPress={() =>
                               onOpenFeedbackForm(serverId || "", FormType.LIKE)
                             }
@@ -144,7 +152,7 @@ const ChatMessage = forwardRef<HTMLDivElement, ChatMessageProps>(
                             isIconOnly
                             variant="ghost"
                             className="p-0"
-                            aria-label="Dislike message"
+                            aria-label="Rate message as unhelpful"
                             onPress={() =>
                               onOpenFeedbackForm(
                                 serverId || "",